Rename `read_time.rb` to `read_time_filter.rb`
Rename `read_time.rb` to `read_time_filter.rb` for consistency.
