+++ lib/read_time_filter.rb
