+++ lib/swedbank-pay-design-guide-jekyll-theme.rb
@@ -2,5 +2,5 @@
 
 require_relative 'sidebar'
 require_relative 'field_tag'
-require_relative 'read_time'
+require_relative 'read_time_filter'
 require_relative 'regex_capture_filter'
